move metadata extraction processor to domain
