+++ server/src/domain/metadata/index.ts
@@ -1,2 +1,2 @@
-export * from './geocoding.repository';
+export * from './metadata.repository';
 export * from './metadata.service';
