+++ server/src/infra/infra.module.ts
@@ -6,13 +6,12 @@ import {
   ICommunicationRepository,
   ICryptoRepository,
   IFaceRepository,
-  IGeocodingRepository,
   IJobRepository,
   IKeyRepository,
   ILibraryRepository,
   IMachineLearningRepository,
   IMediaRepository,
-  immichAppConfig,
+  IMetadataRepository,
   IPartnerRepository,
   IPersonRepository,
   ISearchRepository,
@@ -23,6 +22,7 @@ import {
   ITagRepository,
   IUserRepository,
   IUserTokenRepository,
+  immichAppConfig,
 } from '@app/domain';
 import { BullModule } from '@nestjs/bullmq';
 import { Global, Module, Provider } from '@nestjs/common';
@@ -33,20 +33,20 @@ import { databaseConfig } from './database.config';
 import { databaseEntities } from './entities';
 import { bullConfig, bullQueues } from './infra.config';
 import {
+  APIKeyRepository,
   AccessRepository,
   AlbumRepository,
-  APIKeyRepository,
   AssetRepository,
   AuditRepository,
   CommunicationRepository,
   CryptoRepository,
   FaceRepository,
   FilesystemProvider,
-  GeocodingRepository,
   JobRepository,
   LibraryRepository,
   MachineLearningRepository,
   MediaRepository,
+  MetadataRepository,
   PartnerRepository,
   PersonRepository,
   SharedLinkRepository,
@@ -66,7 +66,7 @@ const providers: Provider[] = [
   { provide: ICommunicationRepository, useClass: CommunicationRepository },
   { provide: ICryptoRepository, useClass: CryptoRepository },
   { provide: IFaceRepository, useClass: FaceRepository },
-  { provide: IGeocodingRepository, useClass: GeocodingRepository },
+  { provide: IMetadataRepository, useClass: MetadataRepository },
   { provide: IJobRepository, useClass: JobRepository },
   { provide: ILibraryRepository, useClass: LibraryRepository },
   { provide: IKeyRepository, useClass: APIKeyRepository },
