+++ server/src/infra/repositories/index.ts
@@ -7,11 +7,11 @@ export * from './communication.repository';
 export * from './crypto.repository';
 export * from './face.repository';
 export * from './filesystem.provider';
-export * from './geocoding.repository';
 export * from './job.repository';
 export * from './library.repository';
 export * from './machine-learning.repository';
 export * from './media.repository';
+export * from './metadata.repository';
 export * from './partner.repository';
 export * from './person.repository';
 export * from './shared-link.repository';
