+++ server/src/microservices/microservices.module.ts
@@ -2,10 +2,9 @@ import { DomainModule } from '@app/domain';
 import { InfraModule } from '@app/infra';
 import { Module } from '@nestjs/common';
 import { AppService } from './app.service';
-import { MetadataExtractionProcessor } from './processors/metadata-extraction.processor';
 
 @Module({
   imports: [DomainModule.register({ imports: [InfraModule] })],
-  providers: [MetadataExtractionProcessor, AppService],
+  providers: [AppService],
 })
 export class MicroservicesModule {}
