+++ server/test/repositories/storage.repository.mock.ts
@@ -4,6 +4,8 @@ export const newStorageRepositoryMock = (): jest.Mocked<IStorageRepository> => {
   return {
     createZipStream: jest.fn(),
     createReadStream: jest.fn(),
+    readFile: jest.fn(),
+    writeFile: jest.fn(),
     unlink: jest.fn(),
     unlinkDir: jest.fn().mockResolvedValue(true),
     removeEmptyDirs: jest.fn(),
